internal: move Pin and Unpin to sys
Pin and Unpin implement library-specific behaviour for BPF object
pinning. Move it to package sys since that abstracts away BPF things.
